Move docstring cooking to dotty
