+++ dottydoc/src/dotty/tools/dottydoc/core/UsecasePhase.scala
@@ -11,14 +11,18 @@ import model.factories._
 import dotty.tools.dotc.core.Symbols.Symbol
 
 class UsecasePhase extends DocMiniPhase {
-  private def defdefToDef(d: tpd.DefDef, sym: Symbol)(implicit ctx: Context) = DefImpl(
-    sym,
-    d.name.show.split("\\$").head, // UseCase defs get $pos appended to their names
-    flags(d), path(d.symbol),
-    returnType(d.tpt.tpe),
-    typeParams(d.symbol),
-    paramLists(d.symbol.info)
-  )
+  private def defdefToDef(d: tpd.DefDef, sym: Symbol)(implicit ctx: Context) = {
+    val name = d.name.show.split("\\$").head // UseCase defs get $pos appended to their names
+    DefImpl(
+      sym,
+      name,
+      flags(d),
+      path(d.symbol).init :+ name,
+      returnType(d.tpt.tpe),
+      typeParams(d.symbol),
+      paramLists(d.symbol.info)
+    )
+  }
 
   override def transformDef(implicit ctx: Context) = { case df: DefImpl =>
     ctx.docbase.docstring(df.symbol).flatMap(_.usecases.headOption.map(_.tpdCode)).map(defdefToDef(_, df.symbol)).getOrElse(df)
